[REF] mail: move RecordListInternal to its own file
Before this commit, `record_list.js` holds both the public RecordList and
the RecordListInternal class behind it, while every other model class has
its internals in a file of its own: `record_internal.js`,
`store_internal.js`, `model_internal.js`.

This commit moves the class and the reclist helpers it shares with
RecordList to `record_list_internal.js`, so the relation storage reads
like the rest of the model.
