+++ addons/mail/static/src/model/record_list.js
@@ -1,238 +1,9 @@
 import { isRecord, untrackFunctions } from "./misc";
+import { RecordListInternal } from "./record_list_internal";
 
-import { markRaw, proxy, toRaw } from "@odoo/owl";
+import { proxy, toRaw } from "@odoo/owl";
 
 /** @typedef {import("./record").Record} Record */
-
-/** @param {RecordList} reclist */
-function getInverse(reclist) {
-    return reclist._.owner.Model._.fieldsInverse.get(reclist._.name);
-}
-
-/** @param {RecordList} reclist */
-function getTargetModel(reclist) {
-    return reclist._.owner.Model._.fieldsTargetModel.get(reclist._.name);
-}
-
-/** @param {RecordList} reclist */
-function isComputeField(reclist) {
-    return reclist._.owner.Model._.fieldsCompute.get(reclist._.name);
-}
-
-/** @param {RecordList} reclist */
-function isEager(reclist) {
-    return reclist._.owner.Model._.fieldsEager.get(reclist._.name);
-}
-
-/** @param {RecordList} reclist */
-function setComputeInNeed(reclist) {
-    reclist._.owner._.fieldsComputeInNeed.set(reclist._.name, true);
-}
-
-/** @param {RecordList} reclist */
-function isComputeOnNeed(reclist) {
-    return reclist._.owner._.fieldsComputeOnNeed.get(reclist._.name);
-}
-
-/** @param {RecordList} reclist */
-function computeField(reclist) {
-    reclist._.owner._.compute(reclist._.owner, reclist._.name, { fromInNeed: true });
-}
-
-/** @param {RecordList} reclist */
-function isOne(reclist) {
-    return reclist._.owner.Model._.fieldsOne.get(reclist._.name);
-}
-
-export class RecordListInternal {
-    /** @type {string} */
-    name;
-    /** @type {Record} */
-    owner;
-    /**
-     * @type {boolean} Technical flag to immediately read attribute.
-     * Useful to read `data` while passing to owl's proxy getter again to register observer.
-     */
-    gettingField = false;
-
-    constructor() {
-        markRaw(this);
-    }
-
-    /**
-     * Version of add() that does not update the inverse.
-     * This is internally called when inserting (with intent to add)
-     * on relational field with inverse, to prevent infinite loops.
-     *
-     * @param {RecordList} recordList
-     * @param {...Record}
-     */
-    addNoinv(recordList, ...records) {
-        const self = this;
-        const store = recordList._store;
-        if (isOne(recordList)) {
-            const last = records.at(-1);
-            if (isRecord(last) && last.in(recordList)) {
-                return;
-            }
-            const record = self.insert(
-                recordList,
-                last,
-                function recordList_AddNoInvOneInsert(record) {
-                    if (record.localId !== recordList.data[0]) {
-                        const old = recordList._proxy.at(-1);
-                        recordList._proxy.data.pop();
-                        old?._.uses.delete(recordList);
-                        recordList._proxy.data.push(record.localId);
-                        self.syncLength(recordList);
-                        record._.uses.add(recordList);
-                    }
-                },
-                { inv: false }
-            );
-            store._.ADD_QUEUE("onAdd", self.owner, self.name, record);
-            return;
-        }
-        for (const val of records) {
-            if (isRecord(val) && val.in(recordList)) {
-                continue;
-            }
-            const record = self.insert(
-                recordList,
-                val,
-                function recordList_AddNoInvManyInsert(record) {
-                    if (recordList.data.indexOf(record.localId) === -1) {
-                        recordList._proxy.data.push(record.localId);
-                        self.syncLength(recordList);
-                        record._.uses.add(recordList);
-                    }
-                },
-                { inv: false }
-            );
-            store._.ADD_QUEUE("onAdd", self.owner, self.name, record);
-        }
-    }
-    /** @param {R[]|any[]} data */
-    assign(recordList, data) {
-        const self = this;
-        const store = recordList._store;
-        return store.MAKE_UPDATE(function recordListAssign() {
-            /** @type {Record[]|Set<Record>|RecordList<Record|any[]>} */
-            const collection = isRecord(data) ? [data] : data;
-            // data and collection could be same record list,
-            // save before clear to not push mutated recordlist that is empty
-            const vals = [...collection];
-            const oldRecords = recordList._proxyInternal.slice
-                .call(recordList._proxy)
-                .map((recordProxy) => toRaw(recordProxy)._raw);
-            const newRecords = vals.map((val) =>
-                self.insert(recordList, val, function recordListAssignInsert(record) {
-                    if (record.notIn(oldRecords)) {
-                        record._.uses.add(recordList);
-                        store._.ADD_QUEUE("onAdd", self.owner, self.name, record);
-                    }
-                })
-            );
-            const inverse = getInverse(recordList);
-            for (const oldRecord of oldRecords) {
-                if (oldRecord.notIn(newRecords)) {
-                    oldRecord._.uses.delete(recordList);
-                    store._.ADD_QUEUE("onDelete", self.owner, self.name, oldRecord);
-                    if (inverse) {
-                        store._.updateFields(oldRecord, {
-                            [inverse]: [["DELETE", self.owner]],
-                        });
-                    }
-                }
-            }
-            recordList._proxy.data = newRecords.map((newRecord) => newRecord.localId);
-            recordList._.syncLength(recordList);
-        });
-    }
-    /**
-     * Version of delete() that does not update the inverse.
-     * This is internally called when inserting (with intent to delete)
-     * on relational field with inverse, to prevent infinite loops.
-     *
-     * @param {RecordList} recordList
-     * @param {...Record}
-     */
-    deleteNoinv(recordList, ...records) {
-        const self = this;
-        const store = recordList._store;
-        for (const val of records) {
-            const record = this.insert(
-                recordList,
-                val,
-                function recordList_DeleteNoInv_Insert(record) {
-                    const index = recordList.data.indexOf(record.localId);
-                    if (index !== -1) {
-                        recordList.splice.call(recordList._proxy, index, 1);
-                        self.syncLength(recordList);
-                    }
-                },
-                { inv: false }
-            );
-            store._.ADD_QUEUE("onDelete", self.owner, self.name, record);
-        }
-    }
-    /**
-     * @param {RecordList} recordList
-     * @param {R|any} val
-     * @param {(R) => void} [fn] function that is called in-between preinsert and
-     *   insert. Preinsert only inserted what's needed to make record, while
-     *   insert finalize with all remaining data.
-     * @param {boolean} [inv=true] whether the inverse should be added or not.
-     *   It is always added except when during an insert on a relational field,
-     *   in order to avoid infinite loop.
-     * @param {"ADD"|"DELETE} [mode="ADD"] the mode of insert on the relation.
-     *   Important to match the inverse. Most of the time it's "ADD", that is when
-     *   inserting the relation the inverse should be added. Exception when the insert
-     *   comes from deletion, we want to "DELETE".
-     */
-    insert(recordList, val, fn, { inv = true, mode = "ADD" } = {}) {
-        const inverse = getInverse(recordList);
-        const targetModel = getTargetModel(recordList);
-        if (typeof val !== "object") {
-            if (Array.isArray(recordList._store[targetModel].id)) {
-                throw new Error(
-                    `Cannot insert "${val}" on relational field "${recordList._.owner.Model.getName()}/${
-                        recordList._.name
-                    }": target model "${targetModel}" doesn't support single-id data!`
-                );
-            }
-            // single-id data
-            val = { [recordList._store[targetModel].id]: val };
-        }
-        if (inverse && inv) {
-            // special command to call addNoinv/deleteNoInv, to prevent infinite loop
-            const target = isRecord(val) && val._raw === val ? val._proxy : val;
-            target[inverse] = [[mode === "ADD" ? "ADD.noinv" : "DELETE.noinv", recordList._.owner]];
-        }
-        /** @type {R} */
-        let newRecordProxy;
-        if (!isRecord(val)) {
-            newRecordProxy = recordList._store[targetModel].preinsert(val);
-        } else {
-            newRecordProxy = val;
-        }
-        const newRecord = toRaw(newRecordProxy)._raw;
-        fn?.(newRecord);
-        if (!isRecord(val)) {
-            // was preinserted, fully insert now
-            recordList._store[targetModel].insert(val);
-        }
-        return newRecord;
-    }
-    /**
-     * Sync reclist.data length with array length, as to not introduce confusion while debugging
-     *
-     * @param {RecordList} reclist
-     */
-    syncLength(reclist) {
-        reclist.length = reclist.data.length;
-    }
-}
 
 /** * @template {Record} R */
 export class RecordList extends Array {
@@ -273,10 +44,10 @@ export class RecordList extends Array {
                     }
                     return res;
                 }
-                if (isComputeField(recordList) && !isEager(recordList)) {
-                    setComputeInNeed(recordList);
-                    if (isComputeOnNeed(recordList)) {
-                        computeField(recordList);
+                if (recordList._.isComputeField() && !recordList._.isEager()) {
+                    recordList._.setComputeInNeed();
+                    if (recordList._.isComputeOnNeed()) {
+                        recordList._.computeField();
                     }
                 }
                 if (name === "length") {
@@ -319,7 +90,7 @@ export class RecordList extends Array {
                                     recordList._.name,
                                     oldRecord
                                 );
-                                const inverse = getInverse(recordList);
+                                const inverse = recordList._.getInverse();
                                 if (inverse) {
                                     store._.updateFields(oldRecord, {
                                         [inverse]: [["DELETE", recordList._.owner]],
@@ -371,7 +142,7 @@ export class RecordList extends Array {
         const recordListFullProxy = this;
         const store = recordList._store;
         return store.MAKE_UPDATE(function recordListPush() {
-            const inverse = getInverse(recordList);
+            const inverse = recordList._.getInverse();
             for (const val of records) {
                 const record = recordList._.insert(
                     recordList,
@@ -420,7 +191,7 @@ export class RecordList extends Array {
             const record = toRaw(recordProxy)._raw;
             record._.uses.delete(recordList);
             store._.ADD_QUEUE("onDelete", recordList._.owner, recordList._.name, record);
-            const inverse = getInverse(recordList);
+            const inverse = recordList._.getInverse();
             if (inverse) {
                 store._.updateFields(record, { [inverse]: [["DELETE", recordList._.owner]] });
             }
@@ -433,7 +204,7 @@ export class RecordList extends Array {
         const recordListFullProxy = this;
         const store = recordList._store;
         return store.MAKE_UPDATE(function recordListUnshift() {
-            const inverse = getInverse(recordList);
+            const inverse = recordList._.getInverse();
             for (let i = records.length - 1; i >= 0; i--) {
                 const record = recordList._.insert(recordList, records[i], (record) => {
                     recordList._proxy.data.unshift(record.localId);
@@ -473,7 +244,7 @@ export class RecordList extends Array {
                 deleteCount,
                 ...newRecordsProxy.map((newRecordProxy) => toRaw(newRecordProxy)._raw.localId)
             );
-            if (isOne(recordList) && start === 0 && deleteCount === 1) {
+            if (recordList._.isOne() && start === 0 && deleteCount === 1) {
                 // avoid replacing whole list, to avoid triggering observers too much
                 if (list.length === 0) {
                     recordList._proxy.data.pop();
@@ -484,7 +255,7 @@ export class RecordList extends Array {
                 recordList._proxy.data = list;
             }
             recordList._.syncLength(recordList);
-            const inverse = getInverse(recordList);
+            const inverse = recordList._.getInverse();
             for (const oldRecord of oldRecords) {
                 oldRecord._.uses.delete(recordList);
                 store._.ADD_QUEUE("onDelete", recordList._.owner, recordList._.name, oldRecord);
@@ -529,7 +300,7 @@ export class RecordList extends Array {
         const recordList = toRaw(this)._raw;
         const store = recordList._store;
         return store.MAKE_UPDATE(function recordListAdd() {
-            if (isOne(recordList)) {
+            if (recordList._.isOne()) {
                 const last = records.at(-1);
                 if (isRecord(last) && recordList.data.includes(toRaw(last)._raw.localId)) {
                     return last;
@@ -617,4 +388,3 @@ untrackFunctions(RecordList.prototype, [
     "splice",
     "unshift",
 ]);
-untrackFunctions(RecordListInternal.prototype, ["assign", "proxySet"]);
